aya_vision_demo/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
cohere>=5.5.0
Werkzeug==2.3.7
Flask-WTF==1.1.1
pytest==7.4.2
requests==2.31.0
Pillow==10.0.1
gunicorn==21.2.0
watchdog==2.3.1
